java: 重み付き抽選 完成

+++ selfjavagame/src/trrne/Vector2.java
@@ -6,14 +6,14 @@ public class Vector2 {
     public static final Vector2 Zero = new Vector2(0, 0);
     public static final Vector2 One = new Vector2(1, 1);
 
-    public float x, y;
+    public double x, y;
 
     public Vector2() {
         x = 0;
         y = 0;
     }
 
-    public Vector2(float $x, float $y) {
+    public Vector2(double $x, double $y) {
         x = $x;
         y = $y;
     }
@@ -26,7 +26,7 @@ public class Vector2 {
         return new Vector2(x - $a.x, y - $a.y);
     }
 
-    public Vector2 div(final float $a) {
+    public Vector2 div(final double $a) {
         return new Vector2(x / $a, y / $a);
     }
 
@@ -34,23 +34,23 @@ public class Vector2 {
         return new Vector2(x * $a.x, y * $a.y);
     }
 
-    public Vector2 mul(final float $a) {
+    public Vector2 mul(final double $a) {
         return new Vector2(x * $a, y * $a);
     }
 
-    public static float cross(final Vector2 $a, final Vector2 $b) {
+    public static double cross(final Vector2 $a, final Vector2 $b) {
         return $a.x * $b.y - $a.y * $b.x;
     }
 
-    public static float dot(final Vector2 $a, final Vector2 $b) {
+    public static double dot(final Vector2 $a, final Vector2 $b) {
         return $a.x * $b.x + $a.y * $b.y;
     }
 
-    public float magnitude() {
+    public double magnitude() {
         return Mathf.sqrtf(x * x + y * y);
     }
 
-    public static float magnitude(final Vector2 $a) {
+    public static double magnitude(final Vector2 $a) {
         return $a.magnitude();
     }
 
@@ -62,19 +62,19 @@ public class Vector2 {
         return $a.normalize();
     }
 
-    public static float distance(final Vector2 $a, final Vector2 $b) {
+    public static double distance(final Vector2 $a, final Vector2 $b) {
         return magnitude($a.sub($b));
     }
 
-    public static float angle(final Vector2 $a, final Vector2 $b) {
-        final float a = $a.magnitude(), b = $b.magnitude();
+    public static double angle(final Vector2 $a, final Vector2 $b) {
+        final double a = $a.magnitude(), b = $b.magnitude();
         if (Math.abs(a) <= 1e-45 || Math.abs(b) <= 1e-45) {
             return 0;
         }
         return Mathf.acosf(dot($a, $b) / a / b) * Mathf.rad2deg;
     }
 
-    public static boolean twins(final Vector2 $v1, final Vector2 $v2, final float tolerance) {
+    public static boolean twins(final Vector2 $v1, final Vector2 $v2, final double tolerance) {
         return Math.abs($v1.x - $v2.x) < tolerance && Math.abs($v1.y - $v2.y) < tolerance;
     }
